finch_ai/requirements.txt
# FastAPI & Web
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# AI & ML
openai==1.10.0
google-generativeai==0.3.2
anthropic==0.8.1
groq==0.11.0
tiktoken==0.5.2

# Search & Metadata
opensearch-py==2.4.2

# Data Generation
faker==22.6.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
httpx==0.26.0
tenacity==8.2.3

# Security
slowapi==0.1.9
